requirements.txt
faiss-cpu
openai
python-dotenv
streamlit
chainlit
tqdm
numpy
